renaming TransitionPoint to Transition;
