+++ src/Transition.cpp
@@ -1,4 +1,4 @@
-#include "TransitionPoint.h"
+#include "Transition.h"
 
 namespace nl = nlohmann;
